[MXML] Support for untyped MXML properties (u:someProperty)
To avoid that using one untyped property leads to any property
being allowed, untyped properties are now collected into an
ad-hoc type, which is then added to the type assertions.
Since that makes all exceptions explicit, the (upcoming)
object literal type checker can still check all other properties
and make sure no additional properties exist.
For TypeScript, the ad-hoc type is converted to a TypeScript
object literal typed with untyped properties, which simply
look like this: { <property>(, <property>)* }
Like in object literal, properties that are not identifiers
need to be quoted.

Also changed that a config object type cast is replaced by a
type assertion not only if the result is handed into a constructor,
but also if it is used for the super constructor call.
Since type assertions generate no code, this simplifies the
generated constructor code of many MXML classes.