[MXML] Type-safe exml:mode="append"/"prepend" support
Use Exml.append()/prepend() instead of an additional ...$at property.
The Exml helper class in ext-as has been adapted accordingly.
It now includes the "at" index in the array object directly.
This is easier to read in generated code, the Exml implementation is
a bit simpler, and it hides any artificial properties from the
generated code which would be hard to type in TypeScript.