[MXML] Fix: report correct error position on error in imported class
When the MXML compiler tries to access another MXML compilation
unit and that other unit contains an error, that error was
incorrectly reported (only) at the current position in the
current compilation unit. This is very confusing for the developer,
since the error description does not match the error position.
Now, both errors are reported, the "root cause" and the error in
the current compilation unit.