[TS] Improve [ArrayElementType] detection
Also detect and use [ArrayElementType] for function
return type.
To achieve this, also find enclosing function declarations
(not the direct parent like for variable declarations, because
the TypeRelation is inside the FunctionExpr of the
FunctionDeclaration) and, after determining the ExpressionType,
check for FunctionSignature and only use its TypeParameter
(see inline comment).