[TS] Config class instead of factory function
Switched from Config type + factory function to Config
class, which, like the main class, must be defined as
a class plus an interface (TypeScript "declaration merging")
to avoid repeating all members of the interface(s).
When creating a Config object, instead of applying the
factory function, the constructor must now be called
('new').