Implements JOO-152 Lazy Initialization Annotation for Singletons
https://jangaroo.myjetbrains.com/youtrack/issue/JOO-152

This is the compiler companion task for LIBS-88.
The new [Lazy] annotation must be detected in non-class primary
declarations and lead to use __lazyFactory__ instead of the
backwards-compatible, non-lazy __factory__.

The annotation is not relevant at run-time
(ANNOTATIONS_FOR_COMPILER_ONLY). An existing test has been
extended to use the new feature and check the expected new
JS code being generated.