Fix incorrect type errors for mixed type accessors
If get and set accessor have mixed type, the type of a dot
expression is different depending on whether it is  used
as the left-hand side (lhs) expression of an assignment.
While there was code to handle this case, it did not always
work. Actually, there were two bugs:
1. When an IdeExpr is "normalized" to a DotExpr, the
   AssignmentOpExpr's first argument is still the IdeExpr,
   not its normalizedExpr.
2. Even when trying to check for the normalizedExpr, too,
   it did not work, because in IdeExpr, DotExpr#setOriginal()
   was called _before_ setting the normalizedExpr.
After fixing both bugs, it now works as expected: The type
of the set accessor is used for lhs expressions, the type of
the get accessor otherwise.
Tested this by changing test class PropertiesAccessBase to
feature a mixed get/set type property and set it with a type
that is not compatible with its get type (Number vs. String).