[MXML] Fix: Add type (xclass) to super constructor call argument
When constructing a config object to hand over to the super constructor,
it seemed unnecessary to use a type cast that would add an xclass
attribute, because the xclass attribute is mainly needed to dispatch
to the right class when instantiating a config object.
So I chose to replace the type cast with a type assertion.
However, it turned out that on rare occasions, the config object is
type-checked. When additionally, the handed-in config that is applied
on the defaults contains a more general xclass, this type check fails.
One (the only?) pattern in Studio workspace that triggered this bug
is an Action config util that creates a default config object with
xclass = "Ext.Action". The config created from MXML is applied on
this config, so if the MXML config has no xclass at all, the xclass
"Ext.Action" will remain. Then, the base class constructor tries to
cast the config object to some special Ext.Action subclass and fails.

The fix is to always set the class name of the current MXML class as
default for xclass. Note that this is a special case, since compared
to other MXML elements, the root element specifies the name of the
superclass, which in this case must be replaced by the name of the
current MXML class to end up with the desired xclass.