[CompilerHint] First "compiler hint" to complement parameter default values
Adding a "compiler hint" to a parameter initializer value like the example
in Function.as adds that default value as an additional argument in the generated
JavaScript / TypeScript to every call where that argument is missing.
This is needed for methods where the AS3 API is rather lenient and allows to
leave out arguments which the TypeScript API doesn't.
The hint is also effective when generating JavaScript code, which should do no
harm, only make the default value explicit.
Compiler hints are specified by annotations. Since there are no parameter
annotations in ActionScript, the [Parameter] annotation is attached to the
method and uses the parameter name. 'required' is the first supported
annotation property, but others will follow.