[JOOC] Improve: Set type of all literal expressions
The only remaining missing type is that of the 'null' literal.
It now leads to type being null, which is a synonym for "any".
Instead, 'null' should not be assignable to primitive types,
this at least gives a warning in IDEA ("Expression type any
is implicitly coerced to unrelated type int/uint/Boolean/...").