[CompilerHint] [Parameter("<name>", coerceTo="<built-in-type>"] to auto-coerce arguments
Reworked and extended the first compiler hint annotation
  [Parameter("<name>", required]
(adding explicit default parameter values when required).
It now also supports coercing arguments for parameters that
have a too-lenient type in AS3 to the (built-in) type they are supposed
to have:
  [Parameter("<name>", coerceTo="<built-in-type>"]
Examples where this annotation is applied in jangaroo-libs are built-in
JavaScript and browser APIs (jangaroo-browser) that allow any type of
value to be handed in, but coerce that value e.g. to a String.
The new compiler hint option primarily aims at producing TypeScript code
matching the standard typings for JavaScript / DOM API, but is also
applied to JavaScript output, so that values are explicitly coerced
before calling the API function instead of relying on the API function
to perform the coercion. This way, we can make sure the new semantics
does not change the behavior before even starting TypeScript migration.