[CompilerHint] [Return("<type>"] to replace return type
The special TypeScript return type "this" is perfect to
type Ext's "chainable" methods. In ActionScript, there is
no such thing as using the type of the current class, so
we either used the type of the class defining the method,
or we even chose to return ':void'.
The first simple implementation of [Return] only supports
[Return("this")]. Actually, it just renders the given
string as the TypeScript return type of the annotated
function. This even works for inheritance, so if you
override or implement a method annotated with
[Return("foo")], your method's return type will be
replaced by ':foo'.
Next step would be to even rewrite / complement return
statements when the AS3 type is 'void', but a TypeScript
return type is given. This would allow to generate better
JS and TS while not introducing breaking AS3 API changes.