[CompilerHint] [Return("this")] on :void method complements 'return this'
The compiler now rewrites / complements return statements when
the AS3 type is 'void', but the TypeScript return type "this"
is given. This allows to generate better JS and TS while not
introducing breaking AS3 API changes.
All Ext methods that are "chainable" must now in Ext AS be
annotated with [Return("this")]. If they return the most
general type, they return ':this' in TypeScript and everything
in fine. Even if they so far have been declared to return
':void', nothing needs to be changed, since the missing
'return this' is now complemented during compilation, to JS
as well as to TS.