[MXML] Render exml:mode="append"/"prepend" like before
The attempt to change the Exml runtime to integrate the
'$at' insert position inside the array instead of adding
an auxillary property failed for some strange corner
cases.
It seems Ext sometimes copies config values, a especialy
arrays seem to be cloned on initConfig() so that our
hidden __at__ property was lost.
It is safer to just keep the old format

    { ...
      "plugins": [...],
      "plugins$at": 0
    }

in JavaScript. To hide the technical details for the
ActionScript *and* TypeScript developer, the methods
Exml.append() and prepend() are kept.
In ActionScript, they claim to "instrument" the
array. This is now implemented as code rewriting in
the compiler, so it happens at compile time.
{ plugins: Exml.append([...]) }
becomes
{ plugins$at: Exml.APPEND, plugins:: ([...]) }
(I decided to just keep the unnecessary parenthesis.)
In TypeScript (not yet implemented), this will be
rewritten to
{ ...Exml.append("plugins", [...]) }
and actually happen at runtime. This trick is only
possible in TypeScript / ES5, since ActionScript
does not support the spread operator '...'.
This means that in TypeScript, the signature of
Exml.append() / prepend() is slightly different,
but since code using it is automatically converted,
this works.