[MXML/TS] Render exml:mode="append" as ...Exml.apppend()
The MXML attribute exml:mode="append" is converted to
ActionScript as a call to Exml.append() ("prepend" is
analog):

    <plugins exml:mode="append">...</plugins>

    { plugins: Exml.append([...]) }

In TypeScript, this is now rewritten to

    { ...Exml.append("plugins", [...]) }

and actually happen at runtime. The spread operator '...'
takes care of "inlining" the two-property object result
of Exml.append() into the current object.

This means that in TypeScript, the signature of
Exml.append() / prepend() is slightly different,
but since code using it is automatically converted,
this works.

For the implementation, the code from JsCodeGenerator
had to be pulled up to CodeGeneratorBase, providing a
hook method handleExmlAppendPrepend() and a utility
getArgFQN(DotExpr) to check for fully-qualified names
like net.jangaroo.ext.Exml. On the fly, this utility
is used for more precise detection of Ext/Exml.apply(),
removing a TODO.