[TS] Instead of import * as AS3, import only used identifiers by name
The compiler now keeps track of usages of the built-in identifiers
'is', 'as', 'cast', 'bind', ''mixin', '_', and imports only the used
ones by name:

  import {is, as} from "@jangaroo/joo/AS3";

The only unused import that may still occur is 'cast', since some
redundant casts are suppressed in code emit phase, when it is too
late to also remove the import.
The TypeScript code looks more readable without the 'AS3.' before
every built-in usage. The only possible downside of this approach
is that there might now be name clashes with local variables, but
when programming Ext TS, these few built-in identifiers should
be considered keywords, anyway.