[L10n] Replace resource bundle name by *_properties class at compile time
In Flex, resource bundles, compiled from *.properties
files, are referenced by their name (string) only.
This aims at avoiding having to compile properties
files before everything is "green" in the IDE.
In TypeScript, however, resource bundles will be
normal ES modules that are edited by the developer
directly. Referring to resource bundles by name
actually _reduces_ static typing.
Thus, the compiler now replaces the name reference
by the actual compilation unit reference, leading
to a valid 'requires' / 'import' and providing IDE
support for resource bundle property access in
TypeScript.
To detect that a String parameter is actually a
resource bundle name that refers to a *_properties
class, the compiler now supports the special
name "PropertiesClass" as a 'coerceTo' value of
the [Parameter] annotation.
All IResourceManager methods that receive a
bundle name must be annotated with
[Parameter("...", coerceTo="PropertiesClass")].
Then, on each call, the String argument will be
replaced by the corresponding class reference.
The resource bundle should be "imported" by the
class annotation [ResourceBundle("...")], but
since this has not been enforced in the past,
our code base often lacks this annotation.
The class is now added to Ext 'requires' /
TypeScript-imported automatically, but a
compiler warning is logged that the
[ResourceBundle] annotation is missing.