[JS+TS] Support different compilation unit target name via [Rename("...")] annotation
For JavaScript classes, the well-known [Native] annotation allows
to choose a different JavaScript fully-qualified name than the one
derived from the ActionScript package and class name.
For TypeScript migration, we now want to map to a different name,
but the source class is not [Native], but an actual ActionScript
class meant to generate code.
To achieve this, the compiler now supports a new annotation
  [Rename("<target-name>")]
at class level, where <target-name> is the desired fully-qualified
name of the class (or more precisely "primary declaration") in the
generated JavaScript / TypeScript code. Like when using [Native],
this name is used everywhere this class is referenced, but here,
since the class actually generates code, also for the definition
of the JavaScript / TypeScript class.
We plan to apply this feature to rename the "ext" package to "Ext"
in ext-as for consistency of "native" Ext JS classes and classes
added by ext-as (which so far used lower-case "ext" even in
JavaScript). Also, all other utility classes in ext-as outside of
"Ext" will be moved to jangaroo-runtime (@jangaroo/joo) to
package "joo". This is also reflected in the heuristics to
generate the NPM package / ES6 module name of referenced classes
in TypeScriptCodeGenerator.