[TS] Fix: Suppress Vector.<T> type parameter when used as value
Vector is mapped to TypeScript Array, and Array also supports a
type parameter, but *not* if used as a value.
The usage of Vector as a type should be handled by another code
path.
Even if not, we almost never use Vector in ActionScript, anyway,
because of its cumbersome syntax, so these few spots may be
fixed manually.