[JS] Fix: [Rename] annotation now handled correctly
With JS target, the [Rename] annotation was handled
just like a [Native] annotation, so the 'require'
logic was also applied. Correct behavior is to always
assume 'require' ([Rename] does not disable 'require').
Slight refactoring: why collect all annotations and
only use the first one when there is a utility method
that returns the first annotation with a given name
in the first place?