[TS] Generate type assertion for computed resource bundle key
When IResourceManager#getString() is called with a concrete
bundle (properties class), but a computed resource key,
TypeScript complains that the key is just any string, not
a 'keyof Instance<PropertiesClass>'. The conversion compiler
now generates the corresponding type assertion so that the
code is green in TypeScript.
To that end, the compiler supports a new 'coerceTo'
annotation parameter value 'keyof <bundle-parameter-name>':

  [Parameter("bundle", coerceTo="PropertiesClass")]
  [Parameter("resourceKey", coerceTo="keyof bundle")]
  function getString(bundle: String, resourceKey: String, ...)

Referencing the parameter with the bundle / properties class
allows to define methods that e.g. copy a property from one
resource bundle to another, i.e. receive two "PropertiesClass"
parameters and two corresponding "keyof" parameters.
Note that there is no such method in IResourceManager et al,
but there are utility methods in CoreMedia code that deal
with ResourceBundles, so this case might actually occur.