[TS] Suppress mixin implementing its own virtual mixin interface
Mixin interfaces were introduced in AS3 to work around the fact
that an AS3 class can only extend one class. So for a mixin,
you must define a mixin interface that acts as a proxy to that
mixin and is used by mixin client classes.
For TypeScript, this is no longer needed, since every class
defines an interface automatically. We already suppress
generating TypeScript files from AS3 mixin interface files
and map any usage of the AS3 mixin interface to the corresponding
mixin class.
The only thing missing was that the AS3 mixin class itself
implements its mixin interface. In this special case, the
interface must not be replaced by the class, but suppressed
completely.