[TS] WIP: Convert [Bindable] config instance access to get-method call
Refactor detecting Config type access.
Now, an ExpressionType can be marked as isConfigType, and the
analyze() phase of several AST nodes (including declarations,
which now also have an expression type) include heuristics
formerly located in TypeScriptCodeGenerator to mark their
expression type as a config type.
The code that would take advantage of that knowledge is
in internalVisitDotExpr(), where a config property access
on an instance (non-config-type) would be rewritten to
calling the get method instead. This is still deactivated
(false -> !type.isConfigType()), because the get/set methods
are not yet generated when converting AS3->TS, so the
generated TypeScript code would contain many more errors.