[TS] Getting a [Bindable] generates getConfig()
Instead of accessing the get method directly, we now
generate a getConfig() call, like so:

  expr.bindableCfg => expr.getConfig("bindableCfg")

THe advantage is that the Babel plugin can detect
usages of .getConfig(string) and rewrite this to

  AS3.getBindable(expr,"bindableCfg")

which is again exactly what the AS->JS compiler
produces from 'expr.bindableCfg'.

Next, must implement the same thing for setting
a bindable Config.