Feature: Allow jangaroo-app-overlay without app dependency
Until this change, a jangaroo-app-overlay had to provide a
dependency on a jangaroo-app(-overlay) artifact. This is
(only) used to determine the intersection of transitive
dependencies, so that the build process knows which
packages to bundle in the overlay artifact.
But there is another, maybe better way: In Maven, such
dependencies are meant to be declared
  <scope>provided</scope>
This makes the app dependency obsolete, which is preferable
for independently developed plugins.

Implementation: Just allow no jangaroo-app(-overlay)
dependency to be present. If so, construct a
JangarooAppOverlay with no (null) baseApp.
After checking for a null baseApp in several locations
and then simply skip the step that adds things from
the baseApp, the resulting packaging look like desired
(if all relevant dependencies are marked as provided).
The goal jangaroo:run works out of the box if you
specify jooProxyTargetSpec=/*
However, code has been adapted to prevent another
NPE even in case of misconfiguration.
Not sure about the new jangaroo-apps (plural!)
packaging, but looking at the code, no NPEs should
occur there.