[TS] Only use untyped ["..."] access when target type is not 'any'
Currently, only when the ActionScript type is not '*', dot access
of undeclared members is rewritten to index access (["..."]).
Now that ActionScript 'Object' is also mapped to TypeScript 'any',
it becomes obvious that this depends on the TypeScript type
being 'any', not on the ActionScript type.
Again, the new toolchain JS result gets closer to the original
JS code!