[AST] Fix [ArrayElementType] usage in ExpressionType
...by refactoring FunctionSignature to support accessors
(MethodType.GET/.SET) and by fixing when and how the
ExpressionType of a function / method and a DotExpr is
determined.
There are clear(er) responsibilities now:
Before, accessor FunctionDeclaration already set their
ExpressionType to the type of the property. This lead
to their actual FunctionSignature being lost.
Now, all FunctionDeclarations have their signature as
their ExpressionType. Being an accessor, not a "normal"
function, is now part of FunctionSignature.
The resolution to an accessor's property type is done
on access, i.e. by DotExpr.
Also, [ArrayElementType] is now checked centrally,
by scope.getExpressionType(IdeDeclaration).
Since the left-hand-side detection for determining
whether a dot expression is a read or a write is
now needed in several places, it has been moved
directly to class Ide.