[TS] Setting a [Bindable] generates setConfig()
Instead of accessing the get method directly, we now
generate a setConfig() call, like so:

  expr.bindableCfg = value => expr.setConfig("bindableCfg", value)

The advantage is that the Babel plugin can detect usages
of expr.setConfig(string, value) and rewrite this to

  AS3.setBindable(expr,"bindableCfg", value)

which is again exactly what the AS->JS compiler
produces from 'expr.bindableCfg = value'.

Some configType flag hand-over had to be implemented
to make this work. The code dealing with type, type
relation, expression type, ... is quite a mess :-(