[TS] Separate [Bindable] and [ExtConfig] properties
In TypeScript, there are now separate classes for [Bindable]
and [ExtConfig] properties.
[Bindable] properties are supposed to end up as "real" Ext
configs, declared in Ext.define() in the 'config' property.
[ExtConfig] properties are actually properties of the target
class which can be set directly (no set method involved)
from the config object.
Considering these semantics, the classes to define these
two flavors of "config" properties are called
<class-name>Properties for [ExtConfigs] and
<class-name>Configs for [Bindable]. While this sounds
counter-intuitive, maybe only the ActionScript annotation
names were confusing? [Bindable] is a standard Flex
annotation, and "bindable" in used in the Ext JS documentation
to describe "real" Ext configs, so callings the classes
<class-name>Configs and <class-name>Bindables would be an
alternative.
Another design decision to make: Currently, config properties
are optional ('?') in ...Configs, but required in ...Properties.
Thus, the config type '_' can extend the ...Configs and the
target class can extend the ...Properties class, and only to
reuse ...Properties in the config type, the application of a
utility type is needed (Partial<...Properties>).
But this might be hard to remember for the developer, or even
be a pitfall. It would be simpler to declare everything required
(no '?') and use Partial<...Configs>, too.