[TS] Fix/improve type of var declaration with object literal initializer
Before, since the type of an object literal is 'Object', a variable
declaration like

    var myMap: Object = {};

was translated to TS

    var myMap = {};

TS type inference then given myMap the type '{}' (object without any
properties). Thus, any untyped access would have to be quoted.
Now, this typical situation is detected as a weakly typed string-to-any
map, in TS called a Record:

  var myMap: Record<string,any> = {};