[JS] Fix: Only look for inherited [ArrayElementType]s for class members
The search for declarations was erroneously also performed for
local variables, leading to strange error messages.
For example, a local
    var items: Array = [];
would lead to type errors because in Container subclasses,
the search ended up in the declaration of Container#items,
assuming the local items' array element type must also be
Component.