[TS] Fix: findSenchaPackageName() works with SWCs from TS mode
TypeScriptCodeGenerator#findSenchaPackageName() scans a
ZipFileInputSource which is actually an SWC to find out
its Sencha package name. The old strategy was to look for
a JSON file containing the package name, but this file is
missing if you build the SWC in convertToTypeScript mode.
The new solution is more robust and even faster: The SWC
always contains a path META-INF/maven/<groupId>/<artifactId>.
The Sencha package name is derived by adding these with
a "__" separator.
ZipEntryInputSource#getName() returned empty string for
"folders", this had to be fixed.