[TS] Fix: object literal key rewritten like ide + arguments not allowed in strict mode
These two bugs seem unrelated, but are fixed at the same
code location.
ObjectLiterals contain ObjectFields, whose label is a
String LiteralExpression or an Ide. Such an Ide must not
be rewritten like an Ide in a normal expression. But how
do we distinguish this?
An ObjectField's label is not analyze()d - on purpose,
because this Ide "must not be confused with a normal Ide".
Since analyze() also sets the parent node, a label Ide
has a null parent node. So we do not rewrite Ide-s that
have a null parent node.
However, IdeDeclaration also does not analyze() its Ide,
so this had to be added.
Which leads us to the second bug. 'arguments' must not
be used as a paramter name in strict mode. The code to
ensure this was already there, but assumed 'arguments'
would only be used as a ...rest parameter. But it is
not allowed as any parameter!
The connection between the bugs is that now, 'arguments'
was only rewritten in usages, not in its declaration,
because IdeDeclaration failed to set its Ide's parent
node.