[TS/AST] Fix arguments renaming
Hopefully eventually fix when 'arguments' is renamed to
'arguments$' and when not. To distinguish an explicit
'arguments' parameter declaration from the given, implicit
one, the latter's symbol is now marked as "virtual".
To simplify creating "virtual" JooSymbols, the setVirtual()
method has been replaced by a builder-pattern virtual()
method. This leads to most code changes in this commit.
So if 'arguments' has a virtual symbol, is must not be
renamed. Furthermore, it must not be renamed if the Ide
is the property of a dot expression, because then it was
just resolved to the parameter because it has the same
name.