[TS/JS/AST] TS support for class annotations -> runtime metadata
After moving the list of annotations "for compiler only" to
class Jooc and implementing a utility method of ClassDeclaration
to return the list of its runtime annotations, implemented
rendering of class metadata in a TS-friendly format. There is
a new "virtual" TS @jangaroo/joo/AS3 API 'metadata()' used
like so:

metadata(<target-class>, ["<meta-name>", { <param-key>: <param-value>* }]*);

with the usual special key "" (empty string) for the annotation's
key-less value.
This format must be rewritten by the Babel plugin to the old
JS class metadata format as a property of the Ext.define() object:

    metadata: {"": [(<meta-name", [("<param-key>", <param-value>)*])*]}

Here, the first property "" stands for the whole class, other keys
represent property annotations, which will probably be supported
later.