[TS] Give up on rewriting hoisted functions to arrow functions
Rewriting a hoisted (named) function declaration to a var
with an arrow function may fail at runtime, so we better
leave such function declarations alone.
While the forward reference worked since 'const' was
changed to 'var', at runtime, the 'var' is initialized too
late. A hoisted function is defined "in advance" and forward
to it will work at runtime.