[TS] Fix: Import inherited [ArrayElementType] correctly
Since [ArrayElementType] contributes to the TypeScript Array
type parameter even when inherited from a declaration in some
super class or interface, it can happen that this type is not
used otherwise, leading to a missing import and incorrect
access code ('com_acme_foo_Bar').
Now, the Array element type compilation unit dependency is
determined and added in TypedIdeDeclaration#analyze(), so that
it is already available when code generation starts.
Also, AbstractScope#findArrayElementType() now takes care of
_not_ inheriting from private members.