[TS] Revamp detection of Config type and properties
To determine whether a class uses the Ext Config
system, the AS->TS compiler no longer considers
[ExtConfig] or [Bindable] annotations, but only
looks at the constructor. (ext.Base is an
exception, as it does not declare a config
constructor, but still has a config type.)
The constructor must have some parameter named
"config" of the class type or one of its
direct subclasses (for MXML base classes).
Of such classes, all public fields or accessors
are now treated as [ExtConfig], so the annotation
is now completely ignored like in AS->JS mode.