[AST] Fix: let getModifiers() for interface properties return 'public'
While a PropertyDeclaration "inherits" all annotations
and modifiers from its getter / setter, the logic that
interface members are implicitly public is implemented
in FunctionDeclaration#getModifiers(), so it does not
apply to PropertyDeclaration. Thus, asking an interface
for its members returns PropertyDeclarations that claimed
not to be public!
This is fixed by delegating getModifiers() of a
PropertyDeclaration to its getter or setter (if both
are present they should have the same modifiers).
isPublic() is then correctly derived from getModifiers().