[TS] Change: Remove [Native] interface special case
All compilation units with [Native] annotation without
'require' are again handled the same way. The special
case for interfaces has been removed, so a [Native]
interface, like any other compilation unit, ends up
in the index.d.ts.
The idea behind the special case was to use it for
TypeScript interfaces, i.e. duck typing interfaces
that do not (need to) support 'is' checks. But as it
turns out, [Native(require)] is the right choice for
this use case. While it is meant for compilation
units that are implemented "natively" in JavaScript,
providing no implementation at all for an interface
is exactly the TypeScript semantics: You can use
such interface as a type only, if you try to use it
as a "value" (which you would need for an 'is'-check),
it results in an error at build-time. The only
missing link is that the Jangaroo compiler would not
yet issue this error, but it is only reported later
by Sencha Cmd.