[TS] Fix: imports in non-module compilation units
This should get importing in non-module compilation
units right, i.e. those that end up in index.d.ts.
Found out that the strange functional import syntax
needed so that index.d.ts does not become a module,
too, can be "stored" in a type declaration. This
way, we can treat those imports quite similar as
"normal" ones, assigning a local name for the imported
compilation unit. It makes the generated (API) code
far more readable.
Also, we refrain from using TypeScript's new
'import type' syntax. It is not (yet) handled well
by IDEs (both IntelliJ and VSCode): they never suggest
or generate it, and if an 'import type' is present
because a class has only been used as a type so far,
but now you want to use it as a value, VSCode even
refuses to auto-import it (without the 'type').
But this does not matter, because Babel TypeScript
is capable of removing imports of identifiers that
are used as type only, so the outcome is the same
as if 'import type' had been used.