[TS] Fix: inline import().default again
Assigning to a "local" type didn't work, because that
type was not so "local" after all. It led to duplicate
declaration errors when the same import is needed in
several compilation units in the same namespace.
Since we only compile one unit at a time, it would be
extremely complicated to gather all such imported types
for the whole Maven module / NPM package.
Since 'int' and 'uint' are also modules, fixed that they
are actually replaced by their local name, which, up to
now, always were 'int' and 'uint', but now may be the
ad-hoc import('@jangaroo/joo/AS3/int').default.