[TS] Fix missing config type casts
It is dangerous to leave of type casts that seem to have
no effect, because type-casting a config object may have
the side-effect of adding an xclass / alias / (x)type.
We now do a more thorough check whether the type cast
targets a config type, so that the config object factory
pattern is applied more reliably.

In Studio, this fixes the "Quick Create" dialogs, which
appeared as small empty dialogs because they were missing
their xtype.