[TS/AST] Completely restore former [Bindable] -> Ext config logic
Even for strange edge cases like

    [Bindable]
    internal function get foo() { ... }

the generated TypeScript code now leads to the exact same JS code.
Logic changes:
* [Bindable] get accessor without corresponding set counts as config
* [Bindable] internal (everything but private / protected) too

The last remaining diff is that in AS3->JS, local functions derived
from [Bindable] accessors are named get$foo() / set$foo() and then
mapped to getFoo / setFoo in the Ext define object, while in TypeScript,
their local names are getFoo() / setFoo() and they are mapped
one-to-one. This actually caused one name-clash with a private
method in one case in the whole Studio workspace, but generating
TypeScript code with '$' in accessor names and rewriting that in
Babel would but ugly (though possible). Another option would be to
change local function names of private methods, but that would lead
to a bigger diff between AS->JS and AS->TS->TS.
The best way would probably be to add a Jangaroo compiler warning
that indicates when a private method name hides a generated Ext
config access method. This would have detected HelpIconMixin's
name clash case.