[TS/AST] Fix: Only guess config type flag for exact var name match 'config'
The former heuristic guessed that any variable declaration
name ending with "[Cc]onfig" denotes a config object.
The problem is, if that assumption is wrong, the generated
code won't work, because access to bindable properties is
not rewritten to get/setConfig(). The other way round, the
generated code leads to a TypeScript error (that member
'get/setConfig' is not found), but nevertheless, the
generated JS code works. So we should rather be conservative
in guessing that something has a config type.

In Studio, this bug lead to an error when trying to open
the Dashboard.