[TS] Suppress 'public' modifier
It is the default in TypeScript, anyway, unlike in ActionScript,
where 'internal' (package-private) is the default. Thus, all
Jangaroo AS3 code uses 'public' explicitly, so the argument that
it is better to do so in TypeScript, too, when migrating untyped
JavaScript code, to distinguish "I have not yet thought about
visibility" versus "yes, it is public" does not apply to an
AS3 Jangaroo code base.