[jangaroo:run] Support for additional-packages (plugins)
There is a new configuration option and system property called
'additionalPackagesDir' for goal jangaroo:run. Its help text
describes it like so:
  Set the absolute file path of a folder containing an additional 'packages' folder.
  Use this to complement an application with plugins that are developed in a
  dedicated workspace available under some local file path.
  This is usually the 'target/app' folder of the Jangaroo App  plugin module (the
  one that uses the Maven goal 'package-plugin' to produce the plugin ZIP).

The new AdditionalPackagesFromFolderServlet is always configured for
jangaroo-app(-overlay) mode and uses the additionalPackagesDir to compute
a JSON response for the directory listing when path /additional-packages/
or a sub-path is requested. If no additionalPackagesDir is given, an empty
JSON array is returned to avoid a 404 error in the browser log.