Fix test app build when using provided theme dependencies
Since 2c20431f, provided dependencies are included in the
'requires' list in package.json, but not bundled with a
'package-plugin' zip. Unfortunately, <scope>provided</scope>
had already been used to prevent theme dependencies (needed
for IDE completion in SASS files) from being included in
the 'requires' list. This lead to build failures of test
applications that now 'require' a theme package that cannot
be found, because it was specified as 'provided'.

The first part of the fix is to add provided dependencies
to the test app. While 'provided' means that the packages
need not be bundled because they are already provided by
the application into which this package is to be included,
this cannot be the case for the test app.
But even that was not enough, because Maven does stops
computing transitive dependencies when encountering a
provided dependency. Thus, if module A has a provided
dependency on a theme and module B has a normal dependency
on A, the test app of B cannot be built because the
theme package is missing.

Thus, the second part of the fix is to leave out all
'provided' theme dependencies in the 'requires' list in
package.json in the first place.
There has been logic before to leave out the (non-provided)
theme, because it must end up in 'theme', not in 'requires'.
Unfortunately, by just looking at the dependency, one
cannot tell a theme package from a code package. So the
code uses the existing utility method
getProjectFromDependency() to load the dependent POM and
check the 'packageType' configuration for the value 'theme'.