[TS] Use #private members with @ts-expect-error
Instead of Symbols, we now use #private members everywhere,
i.e. also for statics and methods. While TypeScript 4.0 does
not yet support these cases, the resulting error can be
suppressed using
  //@ts-expect-error 18022
  static #foo():void { ... }
Everything else, including usages and navigation in IDEs
(using TypeScript language service) from usage to definition,
then works fine.
Also, full support for #private member names is already being
implemented in the timeline for TypeScript 4.3 (4.2 is scheduled
for Feb 2021).
@ts-expect-error is used rather than @ts-ignore, since we want
the build to fail if we forget to remove the directive as soon
as TSC supports the feature.
The number 18022 is added as a reference to the TSC error number
and to make it easy to search & replace these lines after the
TSC update.