[TS] Use [Rename] value also for local class name
So far, [Rename] only changed the target file of the
compilation unit and the import statements on access.
In most cases, only the package/namespace/path was
changed, so this was sufficient. But if you also
rename the local name (the part without the package
qualifier), the local name of the class (and all
auxiliary constructs) was still the original
ActionScript name.
This is now fixed.