Let [Native] class name of _properties API class be computed
When the Ext-JavaScript name of a [Native] class is derived
from the ActionScript package and name one-to-one, there is
no need to specify the native name explicitly in the [Native]
annotation.
This even does more harm than good when one tries to [Rename]
a properties class, as [Native("...")] already renames the
class, so both annotations compete and [Rename] would have to
have higher precedence. I think it is better to avoid this
combination, so leaving out the explicit name in [Native]
makes it clear that the name from [Rename] should be used.

Note that combining [Rename] and [Native] still does not yet
work, since the compiler does not yet support it.